Further code refactoring
