+++ packages/acter_notifify/test/mock_data/mock_object.dart
@@ -1,0 +1,13 @@
+//Object Emoji
+enum MockObject {
+  //LIST OF OBJECT EMOJIS
+  news('🚀'),
+  pin('📌'),
+  event('🗓️'),
+  taskList('📋'),
+  taskItem('☑️');
+
+  const MockObject(this.emoji);
+
+  final String emoji;
+}
